requirements.txt
selenium~=4.10.0
pytest~=7.4.0
python-dotenv~=1.0.0
pandas~=2.0.3
pygsheets~=2.0.5
gdown~=4.7.1
requests~=2.31.0
urllib3~=1.26.16